refactor(mobile): extract pure helpers from session route
- Extract SessionErrorBoundary, LearningModeControl, ConfirmationToast
  into session/_components/; extract getConfidenceCopy into
  session/_lib/confidence-copy.ts
- session/index.tsx shrinks from ~1825 to ~1445 LOC (pure move, no logic
  changes); audit item #1 of route-shrink plan
- No bug fixes, no API surface changes, no test changes required
